+++ packages/commercetools/api-client/src/fragments/index.ts
@@ -271,6 +271,10 @@ export const OrderFragment = `
     billingAddress {
       ...DefaultAddress
     }
+    cart {
+      id
+      version
+    }
   }
 `;