+++ packages/commercetools/composables/src/useCheckout/createPlaceOrder.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import { createMyOrderFromCart, createCart } from '@vue-storefront/commercetools-api';
+import { createMyOrderFromCart, deleteCart } from '@vue-storefront/commercetools-api';
 import initFields from './initFields';
 import { loading } from './shared';
 
@@ -9,12 +9,13 @@ const createPlaceOrder = ({ factoryParams, cartFields, setCart }) => async () =>
   const { id, version } = cartFields.cart.value;
 
   const orderResponse = await createMyOrderFromCart({ id, version });
-  const cartResponse = await createCart();
-  setCart(cartResponse.data.cart);
-  initFields(cartResponse.data.cart);
+  const { order } = orderResponse.data;
+
+  const cartResponse = await deleteCart(order.cart.id, order.cart.version);
+  setCart(null);
 
   loading.value.order = false;
-  return orderResponse.data.order;
+  return order;
 };
 
 export default createPlaceOrder;