+++ packages/core/docs/commercetools/changelog.md
@@ -5,6 +5,8 @@
 
 ## 0.2.4
 
+- Remove defaulting for checkout shipping details ([#5026](https://github.com/DivanteLtd/vue-storefront/issues/5026))
+
 ### Changes
 
 - added `getTotalReviews` and `getAverageRating` to `productGetters` ([#4958](https://github.com/DivanteLtd/vue-storefront/issues/4958))